requirements.txt
black
numpy
pylint
PyHamcrest>=2.0.4
pytest>=7.1.3
